+++ packages/babel-preset-gatsby-package/lib/__tests__/index.js
@@ -1,12 +1,20 @@
-const preset = require(`../`)
+const preset = require(`../index.js`)
 
 jest.mock(`../resolver`, () => jest.fn(moduleName => moduleName))
 
-beforeEach(() => {
-  delete process.env.BABEL_ENV
-})
 
 describe(`babel-preset-gatsby-package`, () => {
+  let babelEnv;
+
+  beforeEach(() => {
+    babelEnv = process.env.BABEL_ENV
+    delete process.env.BABEL_ENV;
+  })
+
+  afterEach(() => {
+    process.env.BABEL_ENV = babelEnv
+  })
+
   describe(`in node mode`, () => {
     it(`specifies the proper plugins`, () => {
       const { plugins } = preset()
@@ -29,11 +37,14 @@ describe(`babel-preset-gatsby-package`, () => {
         nodeVersion,
       })
 
-      const [, opts] = presets.find(preset =>
-        [].concat(preset).includes(`@babel/preset-env`)
-      )
+      const [, opts] = presets.find(preset => [].concat(preset).includes(`@babel/preset-env`))
 
       expect(opts.targets.node).toBe(nodeVersion)
+    })
+
+    it(`can enable compilerFlags`, () => {
+      const { plugins } = preset(null, { availableCompilerFlags: [`MAJOR`] })
+      expect(plugins).toMatchSnapshot()
     })
   })
 
@@ -50,6 +61,11 @@ describe(`babel-preset-gatsby-package`, () => {
 
     it(`specifies proper presets for debugging`, () => {
       const { presets } = preset(null, { browser: true, debug: true })
+      expect(presets).toMatchSnapshot()
+    })
+
+    it(`specifies proper presets for esm`, () => {
+      const { presets } = preset(null, { browser: true, esm: true })
       expect(presets).toMatchSnapshot()
     })
   })