+++ packages/gatsby-source-wordpress/src/steps/preview/index.ts
@@ -285,7 +285,7 @@ export const sourcePreview = async ({
     isPreview: true,
   })
 
-  if (`unstable_createNodeManifest` in actions) {
+  if (`unstable_createNodeManifest` in actions && node) {
     const manifestId = node.databaseId + previewData.modified
 
     reporter.info(